add community images
