+++ src/www/components/Community/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Community';
